requirements.txt

streamlit
numpy
pillow
mtcnn
tensorflow-cpu
opencv-python-headless